Graph point processes

PartialRejectionSampling.AbstractGraphPointProcess — Type

AbstractGraphPointProcess{T<:Vector{Float64}} <: AbstractPointProcess{T}

Abstract type encoding point processes defined on graphs

PartialRejectionSampling.GraphNode — Type

GraphNode{T<:Int64} <: AbstractDiscreteWindow{T}

Structure with unique field idx representing the index of the vertex of a graph

Hard core graph

PartialRejectionSampling.HardCoreGraph — Type

HardCoreGraph{T<:Integer} <: AbstractGraphPointProcess{T}

Concrete type representing a point process on the vertices of a graph $=(V, E)$ parametrized by \beta ≥ 0 which characterizes the distribution on the independent sets of graph, where each vertex is present with marginal probability $\frac{\beta}{1+\beta}$.

In other words, it can also be viewed as the product distribution $\operatorname{Bernoulli}(\frac{\beta}{1+\beta})^{\otimes |V|}$ on the vertices of graph conditioned on forming an independent set,

\[ \mathbb{P}\!\left[ \mathcal{X} = X \right] \propto \prod_{x\in X} \frac{\beta}{1+\beta} 1_{X \text{forms an independent set}}.\]

See also

Section 7.2 of Heng Guo , Mark Jerrum , Jingcheng Liu (2019)

Example 4.1 of MoKr20

PRS.HardCorePointProcess, the spatial counterpart of PRS.HardCoreGraph

PartialRejectionSampling.HardCoreGraph — Method

HardCoreGraph(
    graph::LG.SimpleGraph{T},
    β::Real
) where {T<:Integer}

Construct a HardCoreGraph.

PartialRejectionSampling.generate_sample — Method

generate_sample(
    pp::HardCoreGraph{T};
    rng=-1
)::Vector{T} where {T}

Generate an exact sample from the PRS.SinkFreeGraph.

Default sampler is PRS.generate_sample_prs.

PartialRejectionSampling.generate_sample_prs — Method

generate_sample_prs(
    pp::HardCoreGraph{T};
    rng=-1
)::Vector{T} where {T}

Sample from PRS.HardCoreGraph using Partial Rejection Sampling (PRS), see Section 7.2 of Heng Guo , Mark Jerrum , Jingcheng Liu (2019)

See also

Example 4.1 of MoKr20

Rooted spanning forests

PartialRejectionSampling.RootedSpanningForest — Type

RootedSpanningForest{T<:LG.SimpleDiGraph{Int64}} <: AbstractGraphPointProcess{T}

Concrete type reprensenting a point process defined on the edges of a graph characterizing the uniform distribution on the spanning forests of graph rooted at roots.

It can be viewed as a the product distribution of the uniform distribution on the set of neighbors of each vertex conditioned on forming no cycles.

The object has two fields:

graph::LG.SimpleGraph{Int64}

roots::Set{Int64}

See also

Section 4.2 of Heng Guo , Mark Jerrum , Jingcheng Liu (2019)

PartialRejectionSampling.RootedSpanningForest — Method

RootedSpanningForest(
    graph::LG.SimpleGraph{T},
    roots::Union{Nothing,T,AbstractVector{T},AbstractSet{T}}=nothing
) where {T<:Integer}

Construct a RootedSpanningForest model on graph, rooted at roots

PartialRejectionSampling._generate_sample_rooted_spanning_forest_prs — Method

Generate a rooted spanning forest of graph, uniformly at random among all rooted spanning forests rooted at roots, using Partial Rejection Sampling (PRS), see Section 4.2 of Heng Guo , Mark Jerrum , Jingcheng Liu (2019)

PartialRejectionSampling.generate_sample — Method

generate_sample(
    pp::RootedSpanningForest{T};
    rng=-1
)

Generate an exact sample from the PRS.RootedSpanningForest.

Default sampler is PRS.generate_sample_prs.

PartialRejectionSampling.generate_sample_prs — Method

generate_sample_prs(
    pp::RootedSpanningForest{T};
    rng=-1
)::T where {T<:LG.SimpleDiGraph{Int64}}

Generate a rooted spanning forest of pp.graph, uniformly at random among all rooted spanning forests rooted at pp.roots, using Partial Rejection Sampling (PRS), see Section 4.2 of Heng Guo , Mark Jerrum , Jingcheng Liu (2019)

PartialRejectionSampling.random_neighbor_assignment — Method

random_neighbor_assignment(
    graph::LG.SimpleGraph{T},
    roots;
    rng=-1
)::LG.SimpleDiGraph{T} where {T}

Return a oriented subgraph of graph where each vertex except the roots is connected to a unique neighbor, i.e., each vertex has outdegree equal to one.

Sink free graph

PartialRejectionSampling.SinkFreeGraph — Type

SinkFreeGraph{T<:LG.SimpleDiGraph{Int64}} <: AbstractGraphPointProcess{T}

Concrete type representing a point process on the edges of a graph characterizing the uniform distribution on the orientations of the edges conditioned on the absence of sinks.

PartialRejectionSampling.SinkFreeGraph — Method

SinkFreeGraph(graph::LG.SimpleGraph{T}) where {T<:Int}

Construct a SinkFreeGraph.

PartialRejectionSampling.generate_sample — Method

generate_sample(
    pp::SinkFreeGraph{T};
    rng=-1
)::Vector{T} where {T}

Generate an exact sample from the PRS.SinkFreeGraph.

Default sampler is PRS.generate_sample_prs.

PartialRejectionSampling.generate_sample_prs — Method

generate_sample_prs(
    pp::SinkFreeGraph{T};
    rng=-1
)::T where {T}

Generate an orientated version of pp.graph uniformly at random among all possible orientations conditioned on the absence of sinks, using Partial Rejection Sampling (PRS).

See also

Section 4.1 of Heng Guo , Mark Jerrum , Jingcheng Liu (2019).

Ising model

PartialRejectionSampling.Ising — Type

Ising{T<:Int} <: AbstractGraphPointProcess{T}

The Ising model characterizes a point process defined on the vertices of graph $(V, E)$ with joint density proportional to

\[ \mathbb{P}\!\left[ \mathcal{X}=X \right] \propto \prod_{i \in V} \exp(h_i x_i) \prod_{\{i, j\} \in E} \exp(J x_i x_j)\]

where $(h_i)_{V}$ are called magnetization paremeters and $J$ the interaction coefficient ($J \gtreqless 0$ characterizes ferro/antiferro magnetic interactions).

PartialRejectionSampling.Ising — Method

Ising(
    dims::Vector{T},
    periodic::Bool,
    h::Vector{Real},
    J::Real
) where {T<:Int}

Construct Ising on a grid graph with dimension dims with periodic boundary conditions according to periodic, interaction parameter J and magnetization vector h

PartialRejectionSampling.Ising — Method

Ising(
    dims::Vector{T},
    periodic::Bool,
    h::Real,
    J::Real,
) where {T<:Int}

Construct Ising on a grid graph with dimension dims with periodic boundary conditions according to periodic, interaction parameter J and constant magnetization h

PartialRejectionSampling.Ising — Method

Ising(
    graph::LG.SimpleGraph{T},
    h::Vector{Real},
    J::Real
) where {T<:Int}

Construct Ising on graph with interaction parameter J and magnetization vector h

PartialRejectionSampling.Ising — Method

Ising(
    graph::LG.SimpleGraph{T},
    h::Real,
    J::Real
) where {T<:Int}

Construct Ising on graph with interaction parameter J and constant magnetization h

PartialRejectionSampling.generate_sample — Method

generate_sample(
    pp::Ising{T},
    idx::T;
    rng=-1
)::T where {T<:Int}

Generate an exact sample from the marginal distribution of state $i=$ idx of pp.

More specifically,

\[ x_i \sim \operatorname{Bernoulli}_{-1, 1} (\sigma(h_i)),\]

where $\sigma$ denotes the PRS.sigmoid function.

PartialRejectionSampling.generate_sample_prs — Method

generate_sample_prs(pp::Ising; rng=-1)

Generate an exact sample form pp using Partial Rejection Sampling (PRS), see Section 4.2 of Heng Guo , Mark Jerrum , Jingcheng Liu (2019)

Default sampler is PRS.generate_sample_grid_prs

See also

Weiming Feng , Nisheeth K. Vishnoi , Yitong Yin (2019)

Weiming Feng , Heng Guo , Yitong Yin (2019) PRS.generate_sample_gibbs_perfect

PartialRejectionSampling.bayes_filter — Method

bayes_filter(
    ising::Ising{T},
    state::AbstractVector{T},
    i::T,
    R::Set{T};
    rng=-1
)::Bool

This function is used as a subroutine of PRS.generate_sample_gibbs_perfect.

PartialRejectionSampling.generate_sample! — Method

generate_sample!(
    state::AbstractVector{T},
    indices,
    ising::Ising{T};
    rng=-1
) where {T<:Int}

Generate an exact sample from the marginal distribution of each states of the Ising model ising at the prescribed indices.

PartialRejectionSampling.generate_sample_conditional! — Method

generate_sample_conditional!(
    state::AbstractVector{T},
    i::T,
    ising::Ising{T};
    rng=-1
) where {T<:Int}

Generate an exact sample from the conditional distribution of the state $x_i$ given its neighboring states in ising.graph. More specifically,

\[ x_i \mid x_{N(i)} \sim \operatorname{Bernoulli}_{-1, 1} (\sigma(h_i + J \sum_{j \in N(i)} x_j).\]

PartialRejectionSampling.generate_sample_gibbs_perfect — Method

generate_sample_gibbs_perfect(
    ising::Ising{T};
    rng=-1
)::Vector{T} where {T<:Int}

Generate an exact realization of the PRS.Ising model using a tailored implementation of the perfect Gibbs sampler of Weiming Feng , Heng Guo , Yitong Yin (2019).