Pattern free string

PartialRejectionSampling.PatternFreeString — Type

PatternFreeString{T<:String} <: AbstractPointProcess{T}

Struct with fields

alphabet::Vector{Char},

pattern::Regex,

used to generate strings made of characters from alphabet avoiding the prescribed pattern.

PartialRejectionSampling.PatternFreeString — Method

PatternFreeString(alphabet::Vector{String}, pattern::String)

Construct a PRS.PatternFreeString.

using PartialRejectionSampling
PRS.PatternFreeString(['A', 'C', 'G', 'T'], "ATGTA")

# output

PatternFreeString{String}
- alphabet = ['A', 'C', 'G', 'T']
- pattern = ATGTA

PartialRejectionSampling._pattern_free_string_prs — Method

_generate_pattern_free_string_prs(
    rng::Random.AbstractRNG,
    alphabet::Vector{Char},
    pattern::Regex,
    size::Int
)::String

Generate a string uniformly at random among all strings made of characters from alphabet with no occurence of the pattern pattern, using a tailored version of Partial Rejection Sampling (PRS) derived by Juan Gil , Joshua Amaniampong , Jake Wellens (2018)

PartialRejectionSampling.generate_sample — Method

generate_sample(
    [rng=Random.AbstractRNG,]
    pp::PatternFreeString{T},
    size::Int
)::T where {T<:String}

Generate a string uniformly at random among all strings made of characters from pp.alphabet with no occurence of the pattern pp.pattern.

Default sampler is PRS.generate_sample_prs.

PartialRejectionSampling.generate_sample_prs — Method

generate_sample_prs(
    [rng::Random.AbstractRNG,]
    pp::PatternFreeString{T},
    size::Int
)::T where {T<:String}

Generate a string uniformly at random among all strings made of characters from pp.alphabet with no occurence of the pattern pp.pattern, using a tailored version of Partial Rejection Sampling (PRS).

See also

Technical report of Juan Gil , Joshua Amaniampong , Jake Wellens (2018)

using PartialRejectionSampling
pp = PRS.PatternFreeString(['A', 'C', 'G', 'T'], "ATGTA")
PRS.generate_sample_prs(pp, 20)