Pattern free string

PartialRejectionSampling.PatternFreeString — Type

PatternFreeString{T<:String} <: AbstractPointProcess{T}

Container with fields alphabet and pattern used to generate a string made of characters from alphabet avoiding the prescribed pattern.

PartialRejectionSampling.PatternFreeString — Method

PatternFreeString(alphabet::Vector{String}, pattern::String)

Construct a PRS.PatternFreeString.

using PartialRejectionSampling
PRS.PatternFreeString(["A", "C", "G", "T"], "ATGTA")

# output

PatternFreeString{String}
- alphabet = ["A", "C", "G", "T"]
- pattern = ATGTA

PartialRejectionSampling._generate_sample_pattern_free_string_prs — Method

generate_pattern_free_string_prs(
    alphabet::Vector{T},
    pattern::T,
    size::Int;
    rng=-1
)::T where {T<:AbstractString}

Generate a string uniformly at random among all strings made of characters from alphabet with no occurence of the pattern pattern, using a tailored version of Partial Rejection Sampling (PRS) derived by Juan Gil , Joshua Amaniampong , Jake Wellens (2018)

PartialRejectionSampling.find_bad_ranges — Method

find_bad_ranges(
    pattern::T,
    string::T
)::Vector{UnitRange} where {T<:AbstractString}

Identify where pattern occur in string and return the corresponding ranges of indices.

PartialRejectionSampling.find_characters_to_resample — Method

find_characters_to_resample(
    string_vec::Vector{T},
    pattern::T,
    pref_suff::Vector{U}
)::Vector{U} where {T<:String, U<:Int}

Identify the set of events to be resampled as constructed by Algorithm 5 in Heng Guo , Mark Jerrum , Jingcheng Liu (2019) as part of the Partial Rejection Sampling (PRS) method. Return the indices of the variables involved in the corresponding events.

See also

PRS.find_bad_ranges

PartialRejectionSampling.generate_sample! — Method

generate_sample!(
    string_vec::Vector{T},
    indices,
    alphabet::Vector{T};
    rng=-1
) where {T<:AbstractString}

Generate a character uniformly at random from alphabet at positions prescribed by indices in string_vec

PartialRejectionSampling.generate_sample — Method

generate_sample(
    pfs::PatternFreeString{T},
    size::Int;
    rng=-1
)::T where {T<:String}

Generate a string uniformly at random among all strings made of characters from pfs.alphabet with no occurence of the pattern pfs.pattern.

Default sampler is PRS.generate_sample_prs.

PartialRejectionSampling.generate_sample_prs — Method

generate_sample_prs(
    pfs::PatternFreeString{T},
    size::Int;
    rng=-1
)::T where {T<:String}

Generate a string uniformly at random among all strings made of characters from pfs.alphabet with no occurence of the pattern pfs.pattern, using a tailored version of Partial Rejection Sampling (PRS) derived by Juan Gil , Joshua Amaniampong , Jake Wellens (2018).

using PartialRejectionSampling
pfs = PRS.PatternFreeString(["A", "C", "G", "T"], "ATGTA")
PRS.generate_sample_prs(pfs, 20; rng=1)
# output
"TCCAAATCTCCCCTGTCTAT"